fix: use float for TFLOPS instead of truncating to int
Changed argparse type from int to float for:
- flashinfer_mha_decode.py --fp16-tflops
- deepgemm_gemm.py --gpu-tflops

This preserves precision for H100 (989.5 TFLOPs) instead of
truncating to 989, making MFU calculations more accurate.

Impact: Previous truncation inflated MFU by ~0.05% (989 vs 989.5).

+++ kernel_benchmark/deepgemm_gemm.py
@@ -193,6 +193,6 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("-k", type=int, default=1024, help="[m, k] * [k, n]")
     parser.add_argument("-n", type=int, default=1024, help="[m, k] * [k, n]")
-    parser.add_argument("--gpu-tflops", type=int, default=296, help="GPU FP8 TFLOPS")
+    parser.add_argument("--gpu-tflops", type=float, default=296, help="GPU FP8 TFLOPS")
     args = parser.parse_args()
     main(args)
